refactor(wallet): hoist prepareTx as a separate internal util
Also convert SingleAddressWallet to a directory module
