+++ packages/wallet/src/SingleAddressWallet/index.ts
@@ -1,0 +1,1 @@
+export * from './SingleAddressWallet';
